Remove unused utility files and update reservation services
