+++ src/utilies/alertMessages.js
@@ -1,3 +1,16 @@
+export const titles = {
+  roles:
+    {
+      error: "No se pudo realizar la accion",
+      exito: "La accion se realizo correctamente",
+      confirmar: "Confirmar la accion"
+    }
+  ,
+  error: "Error",
+  correct: "Completado",
+  confirm: "Confirmar"
+};
+
 export const messages = {
   roles: {
     exito: "Rol creado exitosamente",
