--- src/utilies/loadAbort.js
@@ -1,4 +1,0 @@
-export const loadAbort = () => {
-  const controller = new AbortController();
-  return controller;
-};
