--- src/utilies/titles.js
@@ -1,12 +1,0 @@
-export const titles = {
-  roles:
-    {
-      error: "No se pudo realizar la accion",
-      exito: "La accion se realizo correctamente",
-      confirmar: "Confirmar la accion"
-    }
-  ,
-  error: "Error",
-  correct: "Completado",
-  confirm: "Confirmar"
-};
